+++ src/pat/date-picker/date-picker.js
@@ -7,12 +7,12 @@ define(
         'pat-base',
         'pikaday',
         'moment',
-        'moment-timezone-data',
         'modernizr'
     ],
-    function(_, Parser, registry, Base, Pikaday, moment, momenttimezone) {
+    function(_, Parser, registry, Base, Pikaday, moment) {
         var parser = new Parser('date-picker');
         parser.addArgument('behavior', 'styled', ['native', 'styled']);
+        parser.addArgument('format', 'YYYY-MM-DD');
         parser.addArgument('week-numbers', [], ['show', 'hide']);
         parser.addArgument('i18n'); // URL pointing to JSON resource with i18n values
         /* JSON format for i18n
@@ -28,7 +28,7 @@ define(
             name: 'date-picker',
             trigger: '.pat-date-picker',
             init: function() {
-                this.options = $.extend(this.options, parser.parse(this.$el));
+                this.options = $.extend(parser.parse(this.$el), this.options);
                 this.polyfill = this.options.behavior === 'native';
                 if (this.polyfill && Modernizr.inputtypes.date) {
                     return;
@@ -36,15 +36,15 @@ define(
                 if (this.$el.attr('type') === 'date') {
                     this.$el.attr('type', 'text');
                 }
+
                 var config = {
                     field: this.$el[0],
-                    minDate: this.$el.attr('min')
-                        ? moment(this.$el.attr('min')).toDate()
-                        : undefined,
-                    maxDate: this.$el.attr('max')
-                        ? moment(this.$el.attr('max')).toDate()
-                        : undefined,
+                    format: this.options.format,
                     showWeekNumber: this.options.weekNumbers === 'show',
+                    toString: function (date, format) {
+                        var date = moment(date).format(format);
+                        return date;
+                    },
                     onSelect: function() {
                         $(this._o.field)
                             .closest('form')
@@ -53,16 +53,29 @@ define(
                         $(this._o.field).trigger('input-change');
                     }
                 };
+
+                if (this.$el.attr('min')) {
+                    config.minDate = moment(this.$el.attr('min')).toDate();
+                }
+                if (this.$el.attr('max')) {
+                    config.maxDate = moment(this.$el.attr('max')).toDate();
+                }
+
                 if (this.options.i18n) {
                     $.getJSON(this.options.i18n, function(data) {
                         config.i18n = data;
-                        new Pikaday(config);
                     });
-                } else {
-                    new Pikaday(config);
                 }
+
+                new Pikaday(config);
                 return this.$el;
+            },
+
+            isodate: function() {
+                var now = new Date();
+                return now.toISOString().substr(0,10);
             }
+
         });
     }
 );